add TaskDataManager
