+++ .gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+TodoList/View/Base.lproj/EntryViewController~.storyboard
+TodoList/View/TodoList/
